Add protocol for chat and system messages

+++ shared/src/protocol/game_sync/unit_sync.rs
@@ -8,7 +8,7 @@ use crate::{behavior::AxialCoordinates, components::{genome::Hybrid, players::Pl
 /// The synchronization of a tile on the map
 pub struct UnitSync {
     pub position: Property<AxialCoordinates>,
-    pub layer: Property<u16>,
+    pub layer: Property<i32>,
 
 	pub player_id: Property<PlayerId>,
